Migrate helper functions to include/ folder
Note: This isn't tested yet. This also adds TODO comments.
